Finish new clean startup for MonoGame
